% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{always}
\alias{always}
\title{The custom constant function}
\usage{
always(v)
}
\arguments{
\item{v}{a value}
}
\value{
a function which returns v
}
\description{
Return a function which always returns a single value
}
\keyword{functional}

