% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{ap}
\alias{ap}
\title{Function application}
\usage{
ap(f, args)
}
\arguments{
\item{f}{a function to apply}

\item{args}{a list of arguments}
}
\value{
v st. ap(f,list(a,b,c)) == f(a,b,c)
}
\description{
Returns the result of substituting args for the arguments of f.
}
\keyword{functional}

