% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{apFrom}
\alias{apFrom}
\title{Apply from a list of transforms}
\usage{
apFrom(f, fs, v)
}
\arguments{
\item{f}{a function}

\item{fs}{a list of functions to cleave over v}

\item{v}{a value}
}
\value{
q like f(fs1(v),fs2(v),...)
}
\description{
Apply f to the results of transforming v with fs
}
\keyword{functions}

