% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{argCompose}
\alias{argCompose}
\title{Argument-wise composition}
\usage{
argCompose(f, fs)
}
\arguments{
\item{f}{a function}

\item{fs}{a set of functions to be applied to the arguments of f}
}
\value{
a function g st. g(a,b,c) -> f(fs[[1]](a),fs[[2]](b), fs[[3]](c))
}
\description{
Compose functions on the argument list of a function f
}
\keyword{functional}

