% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{cap}
\alias{cap}
\title{Curried (left) function application}
\usage{
cap(...)
}
\arguments{
\item{f}{a function to convert to a function needing a list}
}
\value{
g st. cap(f)(list(a,b,c)) == f(a,b,c)
}
\description{
Returns a function waiting for an argument list.
}
\keyword{functional}

