% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{cleave}
\alias{cleave}
\title{The cleave operator}
\usage{
cleave(fs, v)
}
\arguments{
\item{fs}{a list of functions}

\item{v}{a value}
}
\value{
a list of values [f1(v),f2(v)...]
}
\description{
Apply a list of functions to a value, returning a list
}
\keyword{functional}

