% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{curryLeft}
\alias{curryLeft}
\title{Currying on the left.}
\usage{
curryLeft(f)
}
\arguments{
\item{f}{a function}

\item{...}{arguments list to apply on the left}
}
\value{
g st. curryLeft(f)(a,b,c)(q,r,s) == f(a,b,c,q,r,s)
}
\description{
Returns a function curried applied on the left.
}
\keyword{functional}

