% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{curryRight}
\alias{curryRight}
\title{Currying on the right.}
\usage{
curryRight(f)
}
\arguments{
\item{f}{a function}

\item{...}{arguments list to apply on the left}
}
\value{
g st. curryRight(f)(a,b,c)(q,r,s) == f(q,r,s,a,b,c)
}
\description{
Returns a function curried applied on the right.
}
\keyword{functional}

