% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{deal}
\alias{deal}
\title{The deal operator}
\usage{
deal(fs, vs)
}
\arguments{
\item{fs}{a list of functions}

\item{vs}{a list of values}
}
\value{
a list of values [f1(v1),f2(v2)...]
}
\description{
Apply a list of functions to an equally long list of values
}

