% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{filter}
\alias{filter}
\title{filter a list}
\usage{
filter(f, l)
}
\arguments{
\item{f}{a function}

\item{l}{a list or sequence appropriate for map}
}
\value{
a list with all elements for which f(i)==FALSE are removed
}
\description{
return list items for which f is true
}

