% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{fNot}
\alias{fNot}
\title{function level not}
\usage{
fNot(f)
}
\arguments{
\item{f}{a function}
}
\value{
g a function st g(...) = !f(...)
}
\description{
invert the result of a function
}
\keyword{functional}

