% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{fOr}
\alias{fOr}
\title{function-level OR operator}
\usage{
fOr(...)
}
\arguments{
\item{...}{a list of functions}
}
\value{
f st. f(a,b,c) is true when f1(a,b,c) || f2(a,b,c) ...
}
\description{
Return a new function which is true only when all Fs are true
}
\keyword{functional}

