% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{fUnTuple}
\alias{fUnTuple}
\title{convert a function taking a single list to one taking those values as arguments}
\usage{
fUnTuple(f)
}
\arguments{
\item{f}{a function to convert to a function needing a list}
}
\value{
g st. fUnTuple(f)(a,b,c) == f(list(a,b,c))
}
\description{
Returns a function waiting for a list.
}
\keyword{functional}

