% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{hook}
\alias{hook}
\title{J-style hook}
\usage{
hook(f, g)
}
\arguments{
\item{f}{a function}

\item{g}{a function}
}
\value{
h st. h(x) -> f(x,g(x))
}
\description{
apply f to x and g(x)
}
\keyword{functional}

