% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{map}
\alias{map}
\title{Nickname for Map}
\usage{
map(f, ...)
}
\arguments{
\item{f}{a function of N arguments}

\item{...}{N lists}
}
\value{
a list st. map(f,l1,l2,l3) -> list(f(l1[[1]],l2[[1],l3[[1]]... etc
}
\description{
Map f over lists
}

