% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{partialLeft}
\alias{partialLeft}
\title{FUGU - Function(al/ level) programming in R}
\usage{
partialLeft(f, ...)
}
\arguments{
\item{f}{a function}

\item{...}{arguments list to apply on the left}
}
\value{
g st. partialLeft(f,a,b,c)(q,r,s) == f(a,b,c,q,r,s)
}
\description{
Partial Application on the Left
}
\details{
Returns a function partially applied on the left.
}
\keyword{functional}

