% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{partialRight}
\alias{partialRight}
\title{Partial Application on the Right}
\usage{
partialRight(f, ...)
}
\arguments{
\item{f}{a function}

\item{...}{arguments list to apply on the left}
}
\value{
g st. partialRight(f,a,b,c)(q,r,s) == f(q,r,s,a,b,c)
}
\description{
Returns a function partially applied on the right.
}
\keyword{functional}

