% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{permuteArgs}
\alias{permuteArgs}
\title{permute argument positions}
\usage{
permuteArgs(f, ...)
}
\arguments{
\item{f}{a function}

\item{...}{a list of positions}
}
\value{
g st g(a,b,c) is f where each arg has been taken from the
    index at the position passed in.
}
\description{
modify argument list order of a function
}
\keyword{functional}

