% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{reduce}
\alias{reduce}
\title{Nickname for Reduce}
\usage{
reduce(f, x, init, right = FALSE, accumulate = FALSE)
}
\arguments{
\item{f}{(ac,it) -> ac, a function from an accumulating value and a current item to the next value}

\item{init}{an initial value - otherwise lst[[1]] is used}

\item{lst}{a list of items}
}
\description{
Reduce elements of the list with a function
}
\keyword{functional}

