% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{swap}
\alias{swap}
\title{Swap argument positions}
\usage{
swap(f, p1, p2)
}
\arguments{
\item{f}{a function}

\item{p1}{a position or index}

\item{p2}{a position or index}
}
\value{
g st arguments p1 and p2 are swapped
}
\description{
modify argument list order of a function
}
\keyword{functional}

