% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugu.R
\name{tr}
\alias{tr}
\title{Transform an index}
\usage{
tr(o, k, f)
}
\arguments{
\item{o}{an object}

\item{k}{a key}

\item{f}{a function to transform o[[k]]}
}
\value{
o
}
\description{
equivalent to o[[k]] <- f(o[[k]])
}
\keyword{functional}

